+++ jogo/constantes.py
@@ -10,8 +10,8 @@ POSICOES_OCUPADAS = []
 JANELA = pygame.display.set_mode(TAMANHO_JANELA)
 MARGEM_X = TAMANHO_JANELA[0]/2 - LARGURA_MAPA*BLOCO/2
 MARGEM_Y = TAMANHO_JANELA[1]/2 - ALTURA_MAPA*BLOCO/2
-TAMANHO_IMG_TELA_INICIAL = (TAMANHO_JANELA[1]*0.8,TAMANHO_JANELA[1])
-IMG_TELA_INICIAL = pygame.transform.scale(pygame.image.load('jogo/assets/img/tela_inicial.png'),TAMANHO_IMG_TELA_INICIAL)
+TAMANHO_IMG_TELA_INICIAL = (TAMANHO_JANELA[1]*0.6,TAMANHO_JANELA[1])
+IMG_TELA_INICIAL = pygame.transform.scale(pygame.image.load('jogo/assets\img/tela_inicial_quadrada.png'),TAMANHO_IMG_TELA_INICIAL)
 
 #cores
 PRETO = (0,0,0)
